commit 14-02-2026

+++ src/main/java/com/recaudo/api/domain/model/dto/rest_api/UpdateOrdenPerson.java
@@ -6,6 +6,6 @@ import lombok.Data;
 public class UpdateOrdenPerson {
     private Long personId;
     private Long zonaId;
-    private Long orden;
+    private int  orden;
     private boolean status;
 }
